Discussion of Combining Distributed PoW Difficulty Over Multiple Block Chains For Equivalent Security of a Single Chain Block Chain
Abstract
The following text describes a consensus algorithm that is modeled on the blockchain longest chain algorithm. Although it is presented solely with respect to Proof-of-Work (PoW), it is expected to be combined with a version of Proof-of-Stake (PoS) as a complete consensus algorithm. Therefore this consensus algorithm, on its own, does not address all of the attack vectors that the combined consensus algorithm does. Describing this consensus algorithm independently forms the foundation for the combined PoS and PoW consensus algorithm.